+++ src/common_clj/component/kafka/consumer.clj
@@ -5,9 +5,12 @@
             [plumbing.core :as plumbing]
             [com.stuartsierra.component :as component]
             [common-clj.component.kafka.models :as component.kafka.models]
+            [common-clj.component.kafka.producer :as component.kafka.producer]
+            [camel-snake-kebab.core :as camel-snake-kebab]
             [io.pedestal.interceptor :as interceptor]
             [io.pedestal.interceptor.chain :as chain]
             [taoensso.timbre :as timbre]
+            [common-clj.component.kafka.adapters :as component.kafka.adapters]
             [clojure.tools.logging :as log])
   (:import (org.apache.kafka.clients.consumer KafkaConsumer)
            (java.time Duration)
@@ -26,32 +29,39 @@
               (.subscribe kafka-client topics)
               context)}))
 
-(s/defn kafka-record->clj-message :- component.kafka.models/KafkaMessage
-  [record]
-  (let [message (json/decode (.value record) true)]
-    {:topic (keyword (.topic record))
-     :data  {:payload (:payload message)}}))
-
 (s/defn handler-by-topic
   [topic :- s/Keyword
    topic-consumers]
   (topic topic-consumers))
 
+(s/defn replay-dead-letter!
+  [{:keys [topic data]} :- component.kafka.models/KafkaMessage
+   service-name
+   exception-info :- s/Str
+   producer]
+  (component.kafka.producer/produce! {:topic :create-dead-letter
+                                      :data  {:payload {:service       (camel-snake-kebab/->SCREAMING_SNAKE_CASE_STRING service-name)
+                                                        :topic         (camel-snake-kebab/->SCREAMING_SNAKE_CASE_STRING topic)
+                                                        :exceptionInfo exception-info
+                                                        :payload       (json/encode (:payload data))}}}
+                                     producer))
+
 (def kafka-consumer!
   (interceptor/interceptor
     {:name  ::kafka-consumer
-     :enter (fn [{:keys [kafka-client topic-consumers components] :as context}]
+     :enter (fn [{:keys [kafka-client topic-consumers components service-name] :as context}]
               (assoc context :loop-consumer (future
                                               (while true
                                                 (let [records (seq (.poll kafka-client (Duration/ofMillis 100)))]
                                                   (doseq [record records]
-                                                    (let [{:keys [topic data]} (kafka-record->clj-message record)
+                                                    (let [{:keys [topic data] :as clj-message} (component.kafka.adapters/kafka-record->clj-message record)
                                                           {:keys [handler schema]} (handler-by-topic topic topic-consumers)]
                                                       (try
                                                         (do (s/validate schema (:payload data))
                                                             (handler (:payload data) components))
                                                         (catch Exception e
-                                                          (log/error e))))))))))}))
+                                                          (do (log/error e)
+                                                              (replay-dead-letter! clj-message service-name (str e) (:producer components))))))))))))}))
 
 (s/defrecord Consumer [config datomic producer topic-consumers]
   component/Lifecycle
@@ -69,7 +79,8 @@
           context {:consumer-props  consumer-props
                    :topics          topics
                    :topic-consumers topic-consumers
-                   :components      components}]
+                   :components      components
+                   :service-name    (get-in config [:config :service-name])}]
 
       (when-not topics
         (timbre/error :kafka-topics-not-configured))
@@ -84,10 +95,6 @@
 (defn new-consumer [topic-consumers]
   (map->Consumer {:topic-consumers topic-consumers}))
 
-(defn produced-messages
-  [{:keys [produced-messages]}]
-  (map kafka-record->clj-message @produced-messages))
-
 (defn consumed-messages
   [{:keys [consumed-messages]}]
   @consumed-messages)
@@ -101,35 +108,36 @@
   [topics :- #{s/Keyword}
    produced-messages
    consumed-messages]
-  (-> (filter #(topics (keyword (.topic %))) produced-messages)
+  (-> (filterv #(topics (keyword (.topic %))) produced-messages)
       (clojure.set/difference consumed-messages)))
 
 (defrecord MockKafkaConsumer [config datomic producer topic-consumers]
   component/Lifecycle
 
   (start [this]
-    (let [produced-messages (atom [])
-          consumed-messages (atom [])
-          components (plumbing/assoc-when {}
+    (let [components (plumbing/assoc-when {}
                                           :producer (:producer producer)
                                           :config (:config config)
                                           :datomic (:datomic datomic))
+          consumed-messages (atom [])
+          produced-messages (-> components :producer :produced-messages)
           topics (->> components :config :topics (map keyword) set)
+          service-name (-> components :config :service-name)
           consumer-pool (at-at/mk-pool)]
 
       (at-at/interspaced 100 (fn []
                                (doseq [message-record (messages-that-were-produced-but-not-consumed-yet topics @produced-messages @consumed-messages)]
-                                 (let [{:keys [topic data]} (kafka-record->clj-message message-record)
+                                 (let [{:keys [topic data] :as clj-message} (component.kafka.adapters/kafka-record->clj-message message-record)
                                        {:keys [handler schema]} (handler-by-topic topic topic-consumers)]
                                    (try
                                      (do (s/validate schema (:payload data))
-                                         (handler (:payload data) components)
-                                         (commit-message-as-consumed message-record consumed-messages))
+                                         (handler (:payload data) components))
                                      (catch Exception e
-                                       (log/error e)))))) consumer-pool)
+                                       (do (log/error e)
+                                           (replay-dead-letter! clj-message service-name (str e) (:producer components))))
+                                     (finally (commit-message-as-consumed message-record consumed-messages)))))) consumer-pool)
 
-      (assoc this :consumer {:produced-messages produced-messages
-                             :consumed-messages consumed-messages
+      (assoc this :consumer {:consumed-messages consumed-messages
                              :consumer-pool     consumer-pool})))
 
   (stop [{{:keys [consumer-pool]} :consumer :as this}]