+++ test/unit/common_clj/component/kafka/adapters_test.clj
@@ -1,8 +1,8 @@
-(ns common-clj.component.kafka.consumer-test
+(ns common-clj.component.kafka.adapters-test
   (:require [clojure.test :refer :all]
-            [common-clj.component.kafka.consumer :as component.kafka.consumer]
+            [cheshire.core :as json]
             [matcher-combinators.test :refer [match?]]
-            [cheshire.core :as json])
+            [common-clj.component.kafka.adapters :as component.kafka.adapters])
   (:import (org.apache.kafka.clients.producer ProducerRecord)))
 
 (def kafka-message-record (ProducerRecord. (name :test-topic) (json/encode {:payload {:test-key :test-value}})))
@@ -11,4 +11,4 @@
   (testing "that we can convert a kafka record to a message map"
     (is (match? {:topic :test-topic
                  :data  {:payload {:test-key "test-value"}}}
-                (component.kafka.consumer/kafka-record->clj-message kafka-message-record)))))
+                (component.kafka.adapters/kafka-record->clj-message kafka-message-record)))))